Move plNetObjectDebugger NetClient -> NetCommon.
